test: merge toolkit and helper to single file
