+++ test/helper.js
@@ -4,10 +4,41 @@ const dns = require('node:dns').promises
 const stream = require('node:stream')
 const { promisify } = require('node:util')
 const symbols = require('../lib/symbols')
-const { waitForCb } = require('./toolkit')
 const assert = require('node:assert')
 
 module.exports.sleep = promisify(setTimeout)
+
+function waitForCb (options) {
+  let count = null
+  let done = false
+  let iResolve
+  let iReject
+
+  function stepIn () {
+    if (done) {
+      iReject(new Error('Unexpected done call'))
+      return
+    }
+
+    if (--count) {
+      return
+    }
+
+    done = true
+    iResolve()
+  }
+
+  const patience = new Promise((resolve, reject) => {
+    iResolve = resolve
+    iReject = reject
+  })
+
+  count = options.steps || 1
+  done = false
+
+  return { stepIn, patience }
+}
+module.exports.waitForCb = waitForCb
 
 /**
  * @param method HTTP request method
@@ -493,4 +524,45 @@ module.exports.getServerUrl = function (app) {
   return address === '::1'
     ? `http://[${address}]:${port}`
     : `http://${address}:${port}`
+}
+
+module.exports.partialDeepStrictEqual = function partialDeepStrictEqual (actual, expected) {
+  if (typeof expected !== 'object' || expected === null) {
+    return actual === expected
+  }
+
+  if (typeof actual !== 'object' || actual === null) {
+    return false
+  }
+
+  if (Array.isArray(expected)) {
+    if (!Array.isArray(actual)) return false
+    if (expected.length > actual.length) return false
+
+    for (let i = 0; i < expected.length; i++) {
+      if (!partialDeepStrictEqual(actual[i], expected[i])) {
+        return false
+      }
+    }
+    return true
+  }
+
+  for (const key of Object.keys(expected)) {
+    if (!(key in actual)) return false
+    if (!partialDeepStrictEqual(actual[key], expected[key])) {
+      return false
+    }
+  }
+
+  return true
+}
+
+module.exports.assertNoWarning = function (t) {
+  function doNotWarn () {
+    t.assert.fail('no warning')
+  }
+  process.on('warning', doNotWarn)
+  t.after(() => {
+    process.off('warning', doNotWarn)
+  })
 }
